+++ app/react/Library/components/SortButtons.tsx
@@ -6,60 +6,16 @@ import { t } from 'app/I18N';
 import { wrapDispatch } from 'app/Multireducer';
 import { Icon } from 'UI';
 import { DropdownList } from 'app/Forms';
-import { propertyTypes } from 'shared/propertyTypes';
-import { ObjectIdSchema, PropertySchema } from 'shared/types/commonTypes';
 import { IImmutable } from 'shared/types/Immutable';
 import { ClientTemplateSchema, IStore } from 'app/istore';
 import { omit } from 'lodash';
-
-const isSortableType = (type: PropertySchema['type']) => {
-  switch (type) {
-    case propertyTypes.text:
-    case propertyTypes.date:
-    case propertyTypes.numeric:
-    case propertyTypes.select:
-      return true;
-    default:
-      return false;
-  }
-};
-
-type SortType = {
-  label: string;
-  name: string;
-  value: string;
-  type: string;
-  context?: ObjectIdSchema;
-};
-
-type SearchOptions = {
-  order?: 'asc' | 'desc';
-  sort?: string;
-  searchTerm?: string;
-};
-
-const isSortable = (property: PropertySchema) =>
-  property.filter &&
-  (isSortableType(property.type) || (property.inherit && isSortableType(property.inherit.type!)));
-
-const getSortString = (property: PropertySchema) =>
-  `metadata.${property.name}${property.inherit ? '.inheritedValue' : ''}`;
-
-const getMetadataSorts = (templates: IImmutable<ClientTemplateSchema[]>) =>
-  templates.toJS().reduce((sorts: SortType[], template: ClientTemplateSchema) => {
-    (template.properties || []).forEach((property: PropertySchema) => {
-      if (isSortable(property) && !sorts.find(s => s.name === property.name)) {
-        sorts.push({
-          label: property.label,
-          name: property.name,
-          value: getSortString(property),
-          type: property.type,
-          context: template._id,
-        });
-      }
-    });
-    return sorts;
-  }, []);
+import {
+  getCommonSorts,
+  getMetadataSorts,
+  getPropertySortType,
+  SearchOptions,
+  SortType,
+} from '../helpers/sortComponets';
 
 interface SortButtonsOwnProps {
   stateProperty: string;
@@ -94,30 +50,8 @@ const connector = connect(mapStateToProps, mapDispatchToProps);
 
 type mappedProps = ConnectedProps<typeof connector>;
 
-const getPropertySortType = (selected: SortType): string =>
-  selected.type === 'text' || selected.type === 'select' ? 'string' : 'number';
 const getToggleSearchIcon = (search: SearchOptions) =>
   search.order === 'asc' && search.sort !== '_score' ? 'arrow-up' : 'arrow-down';
-
-const defaultSorts = [
-  { label: 'Title', value: 'title', type: 'text', context: 'System' },
-  { label: 'Date added', value: 'creationDate', type: 'number', context: 'System' },
-  { label: 'Date modified', value: 'editDate', type: 'number', context: 'System' },
-];
-
-const getCommonSorts = (search: SearchOptions) => [
-  ...defaultSorts,
-  ...(search.searchTerm
-    ? [
-        {
-          label: 'Search relevance',
-          value: '_score',
-          type: 'number',
-          context: 'System',
-        },
-      ]
-    : []),
-];
 
 const validateSearch = (search: SearchOptions): SearchOptions =>
   search.sort === '_score' && !search.searchTerm
